Split test into multiple classes
